Change Row and Number dbtype to SMALLINT

+++ web/Shared/Models/Seats/Seat.cs
@@ -3,7 +3,7 @@
     public class Seat
     {
         public int Id { get; set; }
-        public int Row { get; set; }
-        public int Number { get; set; }
+        public short Row { get; set; }
+        public short Number { get; set; }
     }
 }
